+++ src/lib_shell/block_validator.ml
@@ -123,7 +123,7 @@ module Logger =
       let worker_name = "node_block_validator"
     end)
 
-module Worker = Worker.Make (Name) (Event) (Request) (Types) (Logger)
+module Worker = Worker.MakeSingle (Name) (Event) (Request) (Types) (Logger)
 
 type t = Worker.infinite Worker.queue Worker.t