+++ src/lib_shell/peer_validator.ml
@@ -94,7 +94,7 @@ module Logger =
       let worker_name = "node_peer_validator"
     end)
 
-module Worker = Worker.Make (Name) (Event) (Request) (Types) (Logger)
+module Worker = Worker.MakeSingle (Name) (Event) (Request) (Types) (Logger)
 open Types
 
 type t = Worker.dropbox Worker.t