+++ src/lib_workers/worker.ml
@@ -221,14 +221,17 @@ module type T = sig
   val find_opt : 'a table -> Name.t -> 'a t option
 end
 
-module Make
+module Make_internal
     (Name : Worker_intf.NAME)
     (Event : Worker_intf.EVENT)
     (Request : Worker_intf.REQUEST)
     (Types : Worker_intf.TYPES)
     (Logger : Worker_intf.LOGGER
                 with module Event = Event
-                 and type Request.view = Request.view) =
+                 and type Request.view = Request.view)
+    (Worker_events : Worker_events.S
+                       with type view = Request.view
+                        and type critical_error = tztrace) =
 struct
   module Name = Name
   module Event = Event
@@ -508,13 +511,13 @@ struct
 
   let canceler {canceler; _} = canceler
 
-  let lwt_emit w (status : Logger.status) =
+  let lwt_emit w evt =
     let (module LogEvent) = w.logEvent in
     let time = Time.System.now () in
     Lwt.bind
       (LogEvent.emit
          ~section:(Internal_event.Section.make_sanitized Name.base)
-         (fun () -> Time.System.stamp ~time status))
+         (fun () -> Time.System.stamp ~time evt))
       (function
         | Ok () -> Lwt.return_unit
         | Error el ->
@@ -524,7 +527,7 @@ struct
               pp_print_trace
               el)
 
-  let log_event w evt = lwt_emit w (Logger.WorkerEvent (evt, Event.level evt))
+  let log_event w evt = lwt_emit w (evt, Event.level evt)
 
   let record_event w evt = Lwt.ignore_result (log_event w evt)
 
@@ -624,7 +627,8 @@ struct
                     let status = Worker_types.{pushed; treated; completed} in
                     let*! () = Handlers.on_completion w request res status in
                     let*! () =
-                      lwt_emit w (Request (current_request, status, None))
+                      Worker_events.(emit request_no_errors)
+                        (current_request, status)
                     in
                     return_unit)
             | Some u -> (
@@ -645,7 +649,8 @@ struct
                     w.current_request <- None ;
                     let* () = Handlers.on_completion w request res status in
                     let* () =
-                      lwt_emit w (Request (current_request, status, None))
+                      Worker_events.(emit request_no_errors)
+                        (current_request, status)
                     in
                     return (Ok ()))
           in
@@ -667,17 +672,17 @@ struct
               match r with
               | Ok () -> loop ()
               | Error errs ->
-                  let* () = lwt_emit w (Crashed errs) in
+                  let* () = Worker_events.(emit crashed) errs in
                   close handlers w (Some errs)))
     in
     let* r = protect_result ~canceler:w.canceler (fun () -> loop ()) in
     match r with
     | Ok () -> Lwt.return_unit
     | Error Lwt.Canceled | Error Lwt_pipe.Closed | Error Lwt_dropbox.Closed ->
-        let* () = lwt_emit w Terminated in
+        let* () = Worker_events.(emit terminated) () in
         close handlers w None
     | Error exn ->
-        let* () = lwt_emit w (Crashed [Exn exn]) in
+        let* () = Worker_events.(emit crashed) [Exn exn] in
         raise exn
 
   let launch :
@@ -737,8 +742,10 @@ struct
       in
       Nametbl.add table.instances name w ;
       let open Lwt_result_syntax in
-      let started = if id_name = base_name then None else Some name_s in
-      let*! () = lwt_emit w (Started started) in
+      let*! () =
+        if id_name = base_name then Worker_events.(emit started) ()
+        else Worker_events.(emit started_for) name_s
+      in
       let* state = Handlers.on_launch w name parameters in
       w.status <- Running (Time.System.now ()) ;
       w.state <- Some state ;
@@ -756,7 +763,7 @@ struct
        worker ([w.worker]) resolves only once the ongoing request has resolved
        (if any) and some clean-up operations have completed. *)
     let open Lwt_syntax in
-    let* () = lwt_emit w Triggering_shutdown in
+    let* () = Worker_events.(emit triggering_shutdown) () in
     let* () = Error_monad.cancel_with_exceptions w.canceler in
     w.worker
 
@@ -812,4 +819,40 @@ struct
   let () =
     Internal_event.register_section
       (Internal_event.Section.make_sanitized Name.base)
+end
+
+module MakeGroup
+    (Name : Worker_intf.NAME)
+    (Event : Worker_intf.EVENT)
+    (Request : Worker_intf.REQUEST)
+    (Logger : Worker_intf.LOGGER
+                with module Event = Event
+                 and type Request.view = Request.view) =
+struct
+  module Events =
+    Worker_events.Make (Name) (Request)
+      (struct
+        type t = tztrace
+
+        let encoding = Error_monad.trace_encoding
+
+        let pp = Error_monad.pp_print_trace
+      end)
+
+  module MakeWorker (Types : Worker_intf.TYPES) = struct
+    include Make_internal (Name) (Event) (Request) (Types) (Logger) (Events)
+  end
+end
+
+module MakeSingle
+    (Name : Worker_intf.NAME)
+    (Event : Worker_intf.EVENT)
+    (Request : Worker_intf.REQUEST)
+    (Types : Worker_intf.TYPES)
+    (Logger : Worker_intf.LOGGER
+                with module Event = Event
+                 and type Request.view = Request.view) =
+struct
+  module WG = MakeGroup (Name) (Event) (Request) (Logger)
+  include WG.MakeWorker (Types)
 end